requirements/runtime.txt
boto3==1.28.43
botocore==1.31.43
chardet
datasets
fasttext-wheel==0.9.2
hanziconv
httpx==0.27.2
huggingface_hub
jieba
jsonlines
langid
openai==1.56.2
opencv-python
packaging
pandas
Pillow==9.4.0
prettytable
nltk
numpy==1.26.4
pydantic
requests
textstat
toml
transformers
wordninja==2.0.0
zhon